Remove hard-coded application.conf from ConfigFactory.load
This allows the default resolution order to happen so the config can be
overridden via system properties or other means.

For load(), the Config docs state:

Loads a default configuration, equivalent to load(defaultApplication()) in most cases.
This configuration should be used by libraries and frameworks
unless an application provides a different one.